Changes
Update roguebuff.dm

Update roguebuff.dm

+++ code/modules/spells/spell_types/pointed/projectile/unholy_grasp.dm
@@ -3,12 +3,11 @@
 	desc = "Toss forth an unholy snare of blood and guts a short distance, summoned from your leftover trophies sacrificed to Graggar. Like a net, may it snare your target! You will need Viscera to use this."
 	button_icon_state = "unholy_grasp"
 	sound = 'sound/misc/stings/generic.ogg'
-	charge_sound = 'sound/magic/vlightning.ogg'
+	charge_sound = 'sound/magic/charging_lightning.ogg'
 
 	spell_type = SPELL_MIRACLE //it does count as one, funnily enough.
 	antimagic_flags = MAGIC_RESISTANCE_HOLY
 	associated_skill = /datum/skill/magic/holy
-	invocation_type = INVOCATION_NONE //Yeah, Amazing?
 	attunements = list(
 		/datum/attunement/blood = 0.5,
 	)
@@ -21,11 +20,14 @@
 
 /datum/action/cooldown/spell/projectile/blood_net/before_cast()
 	. = ..()
+	if(. & SPELL_CANCEL_CAST)
+		return
 	var/obj/item/held_item = owner.get_active_held_item()
 	if(istype(held_item, /obj/item/alch/viscera))
 		qdel(held_item)
 	else
 		to_chat(owner, "I'm missing viscera to cast this..")
+		reset_spell_cooldown()
 		return . | SPELL_CANCEL_CAST
 
 /obj/projectile/magic/unholy_grasp
@@ -35,8 +37,11 @@
 	range = 3 //Net, So Low range.
 
 /obj/projectile/magic/unholy_grasp/on_hit(atom/hit_atom, datum/thrownthing/throwingdatum)
-	if(..() || !iscarbon(hit_atom))	//if it gets caught or the target can't be cuffed.
-		return	//Abort
+	. = ..()
+	if(. != BULLET_ACT_HIT)
+		return
+	if(!iscarbon(hit_atom))	//if it gets caught or the target can't be cuffed.
+		return
 	ensnare(hit_atom)
 
 /obj/projectile/magic/unholy_grasp/proc/ensnare(mob/living/carbon/C)		//Same code as net but with le flavor.
